requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
bcrypt>=4.0.0
PyJWT>=2.8.0
sqlalchemy>=2.0.0
